requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
brazilcep==7.0.1
catboost==1.2.10
certifi==2026.2.25
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Faker==40.5.1
fonttools==4.61.1
frozenlist==1.8.0
graphviz==0.21
idna==3.11
imbalanced-learn==0.14.1
joblib==1.5.3
kiwisolver==1.4.9
lightgbm==4.6.0
matplotlib==3.10.8
multidict==6.7.1
narwhals==2.17.0
numpy==2.4.2
nvidia-nccl-cu12==2.29.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
plotly==6.5.2
propcache==0.4.1
pyarrow==23.0.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.5
threadpoolctl==3.6.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
xgboost==3.2.0
yarl==1.23.0
streamlit
